requirements.txt
discord.py>=2.3.2,<3.0.0
googletrans==4.0.0-rc1
certifi>=2024.8.30
